Solucion del taller de listas

+++ generido/Producto.java
@@ -1,8 +1,8 @@
 public class Producto {
     private String nombre;
-    private int precio;
+    private double precio;
 
-    public Producto(String nombre, int precio) {
+    public Producto(String nombre, double precio) {
         this.nombre = nombre;
         this.precio = precio;
     }
@@ -11,7 +11,7 @@ public class Producto {
         return nombre;
     }
 
-    public int getPrecio() {
+    public double getPrecio() {
         return precio;
     }
